break external trade into multiple components
